backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==4.0.1",
    "fastapi>=0.124.4",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.9",
    "uvicorn>=0.38.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
